backend/requirements.txt
fastapi
uvicorn
python-dotenv
spacy
supabase
fastcoref
openai
networkx
pydantic
httpx
pytest
hypothesis
pyspellchecker
sentence-transformers
transformers
torch
numpy
pymupdf
scikit-learn
python-multipart
cachetools
